# packages/ivr-migration-api/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IVR Migration API",
    "description": "Manages creation of IVR migration resources.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "def-9": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Unique id of test execution",
            "type": "string"
          },
          "botId": {
            "type": "string"
          },
          "testSetId": {
            "type": "string"
          },
          "testSetName": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "outputDownloadUrl": {
            "description": "The URL to download the test file.",
            "type": "string"
          },
          "failureReasons": {
            "type": "array",
            "items": {
              "description": "Reasons why the test failed",
              "type": "string"
            }
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pending"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "error"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "success"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "stopped"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "in-progress"
                ]
              }
            ]
          }
        },
        "required": [
          "id",
          "botId",
          "testSetId",
          "testSetName",
          "createdAt",
          "createdBy",
          "status"
        ],
        "title": "testExecutionResource"
      },
...
      "def-11": {
        "type": "object",
        "properties": {
          "testSetId": {
            "type": "string"
          }
        },
        "required": [
          "testSetId"
        ],
        "title": "createTestExecutionResource"
      },
...
    }
  },
  "paths": {
...
    "/bots/{botId}/testExecutions": {
      "post": {
        "tags": [
          "TestExecution"
        ],
        "description": "Create a test execution",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-11"
              },
              "examples": {
                "new test execution request": {
                  "value": {
                    "testSetId": "test-set-id"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "botId",
            "required": true,
            "description": "Unique id of bot"
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor"
        ],
        "responses": {
          "201": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/def-9"
                },
                "examples": {
                  "Bot": {
                    "summary": "Existing Test Execution.",
                    "value": {
                      "id": "test-execution-id",
                      "testSetId": "test-set-id",
                      "testSetName": "test-set-name",
                      "botId": "bot-id",
                      "createdAt": "2025-05-30T08:04:07.545Z",
                      "createdBy": "user-id",
                      "updatedAt": "2025-05-30T08:04:07.545Z",
                      "updatedBy": "user-id",
                      "status": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Bad request.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listTestExecutions",
        "summary": "List all test executions.",
        "tags": [
          "TestExecution"
        ],
        "description": "List all test executions.\n\nPermissions:\n- `readers` may list test executions.\n",
        "parameters": [
          {
            "schema": {
              "default": 20,
              "type": "integer"
            },
            "in": "query",
            "name": "count",
            "required": true,
            "description": "Count of results to return, if not specified default to 20"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "token",
            "required": false,
            "description": "Token used to paginate from (exclusive)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "botId",
            "required": true,
            "description": "Unique id of bot"
          },
          {
            "schema": {
              "default": "1.0.0",
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "default": "application/json",
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "x-security-scopes": [
          "admin",
          "contributor",
          "reader"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {
                    "testExecutions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "description": "Unique id of test execution",
                            "type": "string"
                          },
                          "botId": {
                            "type": "string"
                          },
                          "testSetId": {
                            "type": "string"
                          },
                          "testSetName": {
                            "type": "string"
                          },
                          "createdAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "createdBy": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "updatedBy": {
                            "type": "string"
                          },
                          "outputDownloadUrl": {
                            "description": "The URL to download the test file.",
                            "type": "string"
                          },
                          "failureReasons": {
                            "type": "array",
                            "items": {
                              "description": "Reasons why the test failed",
                              "type": "string"
                            }
                          },
                          "status": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "pending"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "error"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "success"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "stopped"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "in-progress"
                                ]
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "botId",
                          "testSetId",
                          "testSetName",
                          "createdAt",
                          "createdBy",
                          "status"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "count": {
                          "type": "number"
                        }
                      }
                    }
                  },
                  "required": [
                    "testExecutions"
                  ]
                },
                "examples": {
                  "List of bots": {
                    "summary": "Paginated list of test execution.",
                    "value": {
                      "testExecutions": [
                        {
                          "id": "test-execution-id",
                          "testSetId": "test-set-id",
                          "testSetName": "test-set-name",
                          "botId": "bot-id",
                          "createdAt": "2025-05-30T08:04:07.545Z",
                          "createdBy": "user-id",
                          "updatedAt": "2025-05-30T08:04:07.545Z",
                          "updatedBy": "user-id",
                          "status": "success"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:30001"
    }
  ],
  "security": [],
...
}